대규모 수정 및 추가
세이브/로드 완료.
플레이어 도끼 공격 안되는 것 수정.
인벤토리나 드랍 창 클릭시 공격 나가는 것 수정.
장비를 안하고 있는데 착용중으로 인식된것 수정.
보스 잡고는 아이템 출현하지 않게 수정.
가방이 가득 찼을 때 경고가 뜨고 습득되지 않게 구현.
아이템 버리기 추가.

+++ Assets/02.Scripts/Common/Quest/QuestData.cs
@@ -10,11 +10,11 @@ public class QuestData
     public int[] npcId;
     public bool isMain;
     public QuestGoal goal;
-    public float expReward;
+    public int expReward;
 
     public bool isActive;
 
-    public QuestData(string name, string content, int[] npc, bool main, int amount, GoalType type, float exp)
+    public QuestData(string name, string content, int[] npc, bool main, int amount, GoalType type, int exp)
     {
         questName = name;
         contentsName = content;
